requirements.txt
fastapi
uvicorn[standard]
torch
numpy
pillow
jinja2
